+++ app/build.gradle
@@ -67,6 +67,10 @@ dependencies {
     implementation 'io.reactivex.rxjava2:rxjava:2.2.20'
     implementation 'io.reactivex.rxjava2:rxandroid:2.1.1'
 
+    // Room
+    implementation "androidx.room:room-runtime:$room_version"
+    kapt "androidx.room:room-compiler:$room_version"
+
     // Naver Maps
     implementation 'com.naver.maps:map-sdk:3.10.2'
     implementation 'com.google.android.gms:play-services-location:18.0.0'
