+++ app/src/main/java/com/mashup/ipdam/ui/search/adapter/PlaceAdapter.kt
@@ -4,9 +4,8 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
-import androidx.recyclerview.widget.RecyclerView
 import com.mashup.ipdam.databinding.ItemPlaceBinding
-import com.mashup.ipdam.ui.search.data.Places
+import com.mashup.ipdam.ui.search.data.entity.kakao.Places
 
 class PlaceAdapter(
     private val placeClickListener: PlaceClickListener
