+++ app/src/main/java/com/mashup/ipdam/ui/search/adapter/PlaceViewHolder.kt
@@ -2,7 +2,7 @@ package com.mashup.ipdam.ui.search.adapter
 
 import androidx.recyclerview.widget.RecyclerView
 import com.mashup.ipdam.databinding.ItemPlaceBinding
-import com.mashup.ipdam.ui.search.data.Places
+import com.mashup.ipdam.ui.search.data.entity.kakao.Places
 
 class PlaceViewHolder(
     val binding: ItemPlaceBinding,
