+++ app/src/main/java/com/mashup/ipdam/ui/search/data/repository/SearchRepository.kt
@@ -1,6 +1,6 @@
 package com.mashup.ipdam.ui.search.data.repository
 
-import com.mashup.ipdam.ui.search.data.PlaceKeyword
+import com.mashup.ipdam.ui.search.data.entity.kakao.PlaceKeyword
 import io.reactivex.Single
 
 interface SearchRepository {
