+++ build.gradle
@@ -38,4 +38,5 @@ ext {
     activity_version = '1.3.0-alpha02'
     hilt_lifecycle_version = '1.0.0-alpha03'
     data_store_version = "1.0.0-alpha06"
+    room_version = "2.2.6"
 }
